pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepsc"
version = "0.1.0"
description = "DeepSC: Deep Foundation Models for Single Cell RNA Sequencing"
authors = [
    { name = "Mou YongLi", email = "mou@dbis.rwth-aachen.de" },
    { name = "Ang Li" },
    { name = "Er Jin" },
    { name = "Sikander Hayat" },
    { name = "Johannes Stegmaier" },
    { name = "Stefan Decker" }
]
license = { text = "MIT" }
readme = "README.md"  # Optional
requires-python = ">=3.10"
urls = { "Homepage" = "https://github.com/MouYongli/DeepSC" }
classifiers = [
    "Development Status :: 1 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
    # Core Deep Learning (GPU versions)
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "einops>=0.7.0",
    "local-attention>=1.9.0",

    # Single-cell Analysis
    "scanpy>=1.10.0",
    "anndata>=0.10.0",
    "cellxgene-census>=1.17.0",

    # Scientific Computing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",

    # Configuration & Experiment Tracking
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "wandb>=0.15.0",

    # Distributed Computing
    "dask>=2023.0.0",

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",

    # Utilities
    "tqdm>=4.65.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",

    # Testing
    "pytest>=7.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["deepsc"]
